requirements.txt
fastapi
uvicorn
sqlalchemy
PyPDF2
langchain
openai
faiss-cpu
python-multipart
pytest
requests
